feat(single-product):added cart redux store
